Restructure helpers to allow them to use when() or using().
The driving force behind this change was the desire to include
optional(), which didn't act like a validator so much as a condition.
The way helpers were structured made it impossible for them to register
conditions, so I had to change that. Now they can register conditions by
simply calling when().