chore(package): require node >= 6
BREAKING CHANGE: This isn't strictly necessary, and for now lgtm will keep working on node v4. However, this removes it as a testing target and the next version will make no guarantees about its compatibility with node v4. Most of the rationale for this comes from dependencies no longer working with node v4.